ClearLoan/backend/requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
